+++ packages/scenes/src/variables/components/VariableValueSelect.tsx
@@ -9,17 +9,43 @@ import { VariableValue, VariableValueSingle } from '../types';
 import { selectors } from '@grafana/e2e-selectors';
 import { GrafanaTheme2, SelectableValue } from '@grafana/data';
 import { css, cx } from '@emotion/css';
+import { getOptionSearcher } from './getOptionSearcher';
+
+const filterNoOp = () => true;
 
 export function VariableValueSelect({ model }: SceneComponentProps<MultiValueVariable>) {
-  const { value, key } = model.useState();
+  const { value, text, key, options, includeAll } = model.useState();
+  const [inputValue, setInputValue] = useState('');
+  const [hasCustomValue, setHasCustomValue] = useState(false);
 
-  const onInputChange = model.onSearchChange
-    ? (value: string, meta: InputActionMeta) => {
-        if (meta.action === 'input-change') {
-          model.onSearchChange!(value);
-        }
+  const optionSearcher = useMemo(
+    () => getOptionSearcher(options, includeAll, value, text),
+    [options, includeAll, value, text]
+  );
+
+  const onInputChange = (value: string, { action }: InputActionMeta) => {
+    if (action === 'input-change') {
+      setInputValue(value);
+      if (model.onSearchChange) {
+        model.onSearchChange!(value);
       }
-    : undefined;
+      return value;
+    }
+
+    return value;
+  };
+
+  const filteredOptions = optionSearcher(inputValue);
+
+  const onOpenMenu = () => {
+    if (hasCustomValue) {
+      setInputValue(String(text));
+    }
+  };
+
+  const onCloseMenu = () => {
+    setInputValue('');
+  };
 
   return (
     <Select<VariableValue>
@@ -27,26 +53,38 @@ export function VariableValueSelect({ model }: SceneComponentProps<MultiValueVar
       placeholder="Select value"
       width="auto"
       value={value}
+      inputValue={inputValue}
       allowCustomValue
       virtualized
+      filterOption={filterNoOp}
       tabSelectsValue={false}
       onInputChange={onInputChange}
-      options={model.getOptionsForSelect()}
+      onOpenMenu={onOpenMenu}
+      onCloseMenu={onCloseMenu}
+      options={filteredOptions}
       data-testid={selectors.pages.Dashboard.SubMenu.submenuItemValueDropDownValueLinkTexts(`${value}`)}
       onChange={(newValue) => {
         model.changeValueTo(newValue.value!, newValue.label!);
+
+        if (hasCustomValue !== newValue.__isNew__) {
+          setHasCustomValue(newValue.__isNew__);
+        }
       }}
     />
   );
 }
 
 export function VariableValueSelectMulti({ model }: SceneComponentProps<MultiValueVariable>) {
-  const { value, key, maxVisibleValues, noValueOnClear } = model.useState();
+  const { value, text, options, key, maxVisibleValues, noValueOnClear, includeAll } = model.useState();
   const arrayValue = useMemo(() => (isArray(value) ? value : [value]), [value]);
-  const options = model.getOptionsForSelect();
   // To not trigger queries on every selection we store this state locally here and only update the variable onBlur
   const [uncommittedValue, setUncommittedValue] = useState(arrayValue);
   const [inputValue, setInputValue] = useState('');
+
+  const optionSearcher = useMemo(
+    () => getOptionSearcher(options, includeAll, value, text),
+    [options, includeAll, value, text]
+  );
 
   // Detect value changes outside
   useEffect(() => {
@@ -71,6 +109,7 @@ export function VariableValueSelectMulti({ model }: SceneComponentProps<MultiVal
   };
 
   const placeholder = options.length > 0 ? 'Select value' : '';
+  const filteredOptions = optionSearcher(inputValue);
 
   return (
     <MultiSelect<VariableValueSingle>
@@ -84,7 +123,7 @@ export function VariableValueSelectMulti({ model }: SceneComponentProps<MultiVal
       tabSelectsValue={false}
       virtualized
       allowCustomValue
-      options={model.getOptionsForSelect()}
+      options={filteredOptions}
       closeMenuOnSelect={false}
       components={{ Option: OptionWithCheckbox }}
       isClearable={true}
@@ -93,6 +132,7 @@ export function VariableValueSelectMulti({ model }: SceneComponentProps<MultiVal
       onBlur={() => {
         model.changeValueTo(uncommittedValue);
       }}
+      filterOption={filterNoOp}
       data-testid={selectors.pages.Dashboard.SubMenu.submenuItemValueDropDownValueLinkTexts(`${uncommittedValue}`)}
       onChange={(newValue, action) => {
         if (action.action === 'clear' && noValueOnClear) {